apps/nlp/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
filelock==3.8.0
grpcio==1.50.0
grpcio-tools==1.50.0
huggingface-hub==0.10.1
idna==3.4
joblib==1.2.0
nltk==3.7
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
protobuf==4.21.9
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tokenizers==0.13.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
urllib3==1.26.12
